fix: debug edge status
