+++ docSite/content/zh-cn/docs/development/upgrading/484.md
@@ -1,0 +1,21 @@
+---
+title: 'V4.8.4'
+description: 'FastGPT V4.8.4 更新说明'
+icon: 'upgrade'
+draft: false
+toc: true
+weight: 821
+---
+
+## 升级指南
+
+- fastgpt 镜像 tag 修改成 v4.8.4
+- fastgpt-sandbox 镜像 tag 修改成 v4.8.4 (选择性，无变更)
+- 商业版镜像 tag 修改成 v4.8.4
+
+## V4.8.4 更新说明
+
+1. 新增 - 应用使用新权限系统。
+2. 优化 - 文本分割增加连续换行、制表符清除，避免大文本性能问题。
+3. 修复 - Debug 模式下，相同 source 和 target 内容，导致连线显示异常。
+4. 调整组件库全局theme。
