+++ projects/app/src/components/core/workflow/Flow/nodes/NodeIfElse/ListItem.tsx
@@ -65,10 +65,7 @@ const ListItem = ({
       >
         <Container w={snapshot.isDragging ? '' : 'full'} className="nodrag">
           <Flex mb={4} alignItems={'center'}>
-            <DragIcon
-              visibility={ifElseList.length > 1 ? 'visible' : 'hidden'}
-              provided={provided}
-            />
+            {ifElseList.length > 1 && <DragIcon provided={provided} />}
             <Box color={'black'} fontSize={'md'} ml={2}>
               {getElseIFLabel(conditionIndex)}
             </Box>
